Move Query into a new query.py file.
