+++ docs/monitoring-query.rst
@@ -1,0 +1,7 @@
+Time Series Query
+=================
+
+.. automodule:: gcloud.monitoring.query
+  :members:
+  :undoc-members:
+  :show-inheritance:
