+++ gcloud/monitoring/__init__.py
@@ -19,11 +19,11 @@ from gcloud.monitoring.connection import Connection
 from gcloud.monitoring.label import LabelDescriptor
 from gcloud.monitoring.metric import Metric
 from gcloud.monitoring.metric import MetricDescriptor
+from gcloud.monitoring.query import Query
 from gcloud.monitoring.resource import Resource
 from gcloud.monitoring.resource import ResourceDescriptor
 from gcloud.monitoring.timeseries import Point
 from gcloud.monitoring.timeseries import TimeSeries
-from gcloud.monitoring.timeseries import Query
 
 
 SCOPE = Connection.SCOPE
