+++ gcloud/monitoring/resource.py
@@ -158,6 +158,8 @@ class ResourceDescriptor(object):
 class Resource(collections.namedtuple('Resource', 'type labels')):
     """A monitored resource identified by specifying values for all labels.
 
+    This is a named tuple. See :func:`collections.namedtuple`.
+
     :type type: string
     :param type: The resource type name.
